+++ id_manager/tests/src/IDTests.cpp
@@ -1,6 +1,6 @@
-#include <essentials/Identifier.h>
-#include <essentials/IDManager.h>
-#include <essentials/WildcardID.h>
+#include <id_manager/Identifier.h>
+#include <id_manager/IDManager.h>
+#include <id_manager/WildcardID.h>
 
 #include <gtest/gtest.h>
 #include <vector>
